move calculateDaysSinceLastModified to utilityfunctions.ts
